Rag_Backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
bcrypt==3.2.0
boto3==1.35.84
botocore==1.35.84
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
deepseek==1.0.0
distro==1.9.0
dnspython==2.7.0
docx2txt==0.9
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.7
fastembed==0.2.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
h2==4.2.0
hiredis==3.2.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
keybert==0.8.5
loguru==0.7.3
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
nltk==3.9.1
numpy==2.3.1
ollama==0.5.1
onnx==1.18.0
onnxruntime==1.22.0
openai==1.91.0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
portalocker==2.10.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.2.3
pypdfium2==4.30.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.9.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==2.7.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel @ git+https://github.com/tiangolo/sqlmodel.git@68fc5fa0fdc0e8745c06de19b6ab6081bd70243b
starlette==0.37.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.39.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1